rename sharded copying manager to copying manager.
