+++ scalyr_agent/copying_manager/__init__.py
