+++ scalyr_agent/copying_manager/common.py
