+++ tests/unit/builtin_monitors/kubernetes_monitor_test.py
@@ -44,13 +44,13 @@ from scalyr_agent.builtin_monitors.kubernetes_monitor import (
     ControlledCacheWarmer,
     _ignore_old_dead_container,
 )
-from scalyr_agent.sharded_copying_manager import CopyingManager
+from scalyr_agent.copying_manager import CopyingManager
 from scalyr_agent.util import FakeClock, FakeClockCounter
 from scalyr_agent.test_base import ScalyrTestCase, BaseScalyrLogCaptureTestCase
 from scalyr_agent.test_util import ScalyrTestUtils
 
 from tests.unit.monitor_utils.k8s_test import FakeCache
-from tests.unit.copying_manager_test import FakeMonitor
+from tests.unit.copying_manager_tests.copying_manager_test import FakeMonitor
 from scalyr_agent.test_base import skipIf
 
 import mock
